requirements.txt
textual
textual-inputs
pynput
loguru
pyfiglet
spotipy
fuzzywuzzy
python-dotenv